refactor: page내 style 파일 생성, 파일 구조 변경
