+++ src/pages/signup/SignUpBasicInformationForm/style.tsx
@@ -1,9 +1,8 @@
-import { useState } from "react";
 import styled from "styled-components";
 
-const SingupContainerDiv = styled.div`
+export const SingupContainerDiv = styled.div`
   width: 48.75rem;
-  height: 34.375rem;
+  height: 47rem;
   flex-shrink: 0;
   justify-content: center;
   align-items: center;
@@ -15,15 +14,15 @@ const SingupContainerDiv = styled.div`
   background: #FFF;
   box-shadow: 0px 2px 6px 0px rgba(0, 0, 0, 0.20);
 `;
-const SignUpTitle = styled.p`
+export const SignUpTitle = styled.p`
   color: #36BF7F;
   font-size: 2rem;
 `;
-const SignUpSubTitle = styled.p`
+export const SignUpSubTitle = styled.p`
   color: #303030;
   font-size: 0.875rem;
 `
-const SignUpModal =  styled.div`
+export const SignUpModal =  styled.div`
   display: flex;
   width: 11.5rem;
   height: 1.5rem;
@@ -31,7 +30,7 @@ const SignUpModal =  styled.div`
   align-items: center;
   flex-shrink: 0;
 `;
-const HorizonLineDiv = styled.div`
+export const HorizonLineDiv = styled.div`
   width: 43.75rem;
 	margin-bottom: 1rem;
 	text-align: center;
@@ -40,21 +39,10 @@ const HorizonLineDiv = styled.div`
 `;
 
 
-
-export default function SignUpCompleteForm() : JSX.Element {
-
-  return (
-    <SingupContainerDiv>
-      <SignUpTitle>회원가입 완료</SignUpTitle>
-      <SignUpSubTitle>
-        위게더의 회원이 되신것을 환영합니다!
-      </SignUpSubTitle>
-      <SignUpModal>모달2</SignUpModal>
-      <HorizonLineDiv/>
-      <div>
-        일러스트 이미지
-      </div>
-    </SingupContainerDiv>
-
-  );
-};
+export const StyledForm = styled.form`
+  justify-content: center;
+  align-items: center;
+  place-content: center;
+  display: flex;
+  flex-direction: column;
+`;
