+++ src/pages/signup/SignUpEmailVerificationForm/index.tsx
@@ -1,0 +1,99 @@
+import LargeSignUpButtonGreen from '@/components/button/LargeSignUpButtonGreen';
+import { useState } from 'react';
+import RegularSignUpButtonGreen from '@/components/button/RegularSignUpButtonGreen';
+import RegularSignUpButtonGray from '@/components/button/RegularSignUpButtonGray';
+import { StyledForm } from '@/styles/login/style';
+import {
+	SignUpWrapper,
+	SignUpP,
+	EmailInputContainer,
+	SignUpInput,
+	EmailButton,
+	NextButtonWrapper,
+  SingupContainerDiv,
+  HorizonLineDiv,
+  SignUpModal,
+  SignUpSubTitle,
+  SignUpTitle,
+} from './style';
+
+export type props = {
+	handleNext: any;
+};
+export default function SignUpEmailVerificationForm({ handleNext }: props): JSX.Element {
+	const [schoolEmail, setSchoolEmail] = useState<string>('');
+	const [verificationCode, setVerificationCode] = useState('');
+
+	const [isSubmitClicked, setIsSubmitClicked] = useState(false);
+
+	const handleVerificationCodeSend = () => {
+		// 이메일 인증 번호를 보내는 로직
+		// ...
+	};
+	const checkVerificationCode = () => {
+		// 인증코드 체크 함수
+	};
+	const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+		event.preventDefault();
+		handleNext();
+	};
+
+	return (
+		<SingupContainerDiv>
+			<SignUpTitle>학교 인증</SignUpTitle>
+			<SignUpSubTitle>이 단계는 건너뛸 수 있어요</SignUpSubTitle>
+			<SignUpModal>모달2</SignUpModal>
+			<HorizonLineDiv />
+			<StyledForm onSubmit={handleSubmit}>
+				<SignUpWrapper>
+					<SignUpP>학교 웹메일</SignUpP>
+					<EmailInputContainer>
+						<SignUpInput
+							type="text"
+							value={schoolEmail}
+							placeholder="학교 이메일을 입력해주세요."
+							onChange={(e: any) => {
+								setSchoolEmail(e.target.value);
+							}}
+							required
+						/>
+						<EmailButton
+							onClick={(e: any) => {
+								e.preventDefault();
+								setIsSubmitClicked(true);
+								handleVerificationCodeSend();
+							}}
+						>
+							인증번호 전송
+						</EmailButton>
+					</EmailInputContainer>
+				</SignUpWrapper>
+				<SignUpWrapper>
+					{isSubmitClicked ? (
+						<>
+							<SignUpP>학교 웹메일</SignUpP>
+							<EmailInputContainer>
+								<SignUpInput
+									type="text"
+									value={verificationCode}
+									placeholder="인증번호를 입력해주세요."
+									onChange={(e: any) => {
+										setVerificationCode(e.target.value);
+									}}
+									required
+								/>
+								<EmailButton onClick={checkVerificationCode}>인증번호 확인</EmailButton>
+							</EmailInputContainer>
+						</>
+					) : (
+						<></>
+					)}
+				</SignUpWrapper>
+				<NextButtonWrapper>
+					<RegularSignUpButtonGray text="건너뛰기" type="submit" />
+					<RegularSignUpButtonGreen text="회원가입" type="submit" />
+				</NextButtonWrapper>
+			</StyledForm>
+		</SingupContainerDiv>
+	);
+}
